requirements.txt
# LLM Providers
anthropic>=0.18.0
mistralai>=0.1.0
google-genai>=1.0.0
groq>=0.4.0
openai>=1.12.0

# LangChain
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0
langchain-groq>=0.1.0
langchain-mistralai>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
pyyaml>=6.0
pydantic>=2.0
pandas>=2.0
numpy>=1.24
scipy>=1.11
yfinance>=0.2.30
rich>=13.0
matplotlib>=3.7
pytest>=7.4
python-dotenv>=1.0
